refactor: Moved methods, updated comments for clarity
