+++ src/components/tabs/scatter_plot_data_utils.ts
@@ -42,11 +42,12 @@ export function getBucketIndex(value: number, minValue: number, maxValue: number
 }
 
 /**
- * Splits a trace into one or more traces so that all traces have at most `maxPoints`
- * data points.
+ * Splits a trace into one or more smaller subtraces so that all the subtraces have at
+ * most `maxPoints` data points.
  * @param traceData The trace to split. (Will not be modified.)
- * @param maxPoints The maximum number of points any trace can have.
- * @returns An array of traces, with the same color as the original trace.
+ * @param maxPoints The maximum number of points any subtrace can have.
+ * @returns An array of subtraces, with the same color as the original trace. Each will have
+ * a (non-overlapping) subset of the original trace's data points.
  */
 export function splitTraceData(traceData: TraceData, maxPoints: number): TraceData[] {
   if (traceData.x.length <= maxPoints) {
@@ -67,12 +68,36 @@ export function splitTraceData(traceData: TraceData, maxPoints: number): TraceDa
   return traces;
 }
 
-/** Draws a simple line graph over the data points. */
+/**
+ * Returns true if a Plotly mouse event took place over a histogram subplot.
+ */
+export function isHistogramEvent(eventData: Plotly.PlotMouseEvent): boolean {
+  return eventData.points.length > 0 && eventData.points[0].data.type === "histogram";
+}
+
+/**
+ * Returns a hex color string with increasing transparency as the number of markers increases.
+ * Base color must be a 7-character RGB hex string (e.g. `#000000`).
+ */
+export function applyMarkerTransparency(numMarkers: number, baseColor: string): string {
+  if (baseColor.length !== 7) {
+    throw new Error("ScatterPlotTab.getMarkerColor: Base color '" + baseColor + "' must be 7-character hex string.");
+  }
+  // Interpolate linearly between 80% and 25% transparency from 0 up to a max of 1000 markers.
+  const opacity = remap(numMarkers, 0, 1000, 0.8, 0.25);
+  return (
+    baseColor +
+    Math.floor(opacity * 255)
+      .toString(16)
+      .padStart(2, "0")
+  );
+}
+
+/** Draws a simple line graph with the provided data points. */
 export function makeLineTrace(xData: DataArray, yData: DataArray): Partial<Plotly.PlotData> {
   return {
     x: xData,
     y: yData,
-    name: "",
     type: "scattergl",
     mode: "lines",
     line: {
@@ -99,7 +124,7 @@ export function drawCrosshair(x: number, y: number): Partial<PlotData>[] {
       symbol: "cross-thin",
     },
   };
-  // Add a transparent white outline around the marker for contrast.
+  // Add a transparent white outline behind the marker for contrast.
   const crosshairBg = { ...crosshair };
   crosshairBg.marker = {
     ...crosshairBg.marker,
